Moving utilities to a more reusable place
