+++ src/tests/utilities.ts
@@ -1,0 +1,19 @@
+import { Team } from '../entities/team';
+import { Judge } from '../entities/judge';
+
+export async function createTeamData(numTeams: number): Promise<Team[]> {
+  const teams: Team[] = [];
+  for (let i = 0; i < numTeams; i += 1) {
+    const team = await new Team(`Team ${i}`, i, `Project for Team ${i}`, [String(i)]).save();
+    teams.push(team);
+  }
+  return teams;
+}
+
+export async function createJudgeData(numJudges: number): Promise<Judge[]> {
+  const judges: Judge[] = [];
+  for (let i = 0; i < numJudges; i += 1) {
+    judges.push(await new Judge().save());
+  }
+  return judges;
+}
